+++ CHANGELOG.md
@@ -1,7 +1,15 @@
-<!-- Keep a Changelog guide -> https://keepachangelog.com -->
-
 # search-string-occurrences-intellij-plugin Changelog
 
-## [Unreleased]
+## [0.1.2] - 2025-11-06
+## Added
+- Adds a feature to search hidden files 
+
+## [0.1.1] - 2025-11-05
+### Added
+- Adds a tool window for concurrent text search
+- Implements concurrent search logic using Kotlin Coroutines
+- Adds real-time results and cancellation button
+
+## [0.1.0] - 2025-11-02
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)